requirements.txt
backcall==0.1.0
certifi==2019.11.28
cffi==1.13.2
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
h5py==2.10.0
ipython==7.10.1
ipython-genutils==0.2.0
jedi==0.15.1
kiwisolver==1.1.0
matplotlib==3.1.2
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.4
olefile==0.46
pandas==0.25.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==3.0.2
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.5
PySCIPOpt==1.4.9
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
scipy==1.3.3
six==1.13.0
torch==1.3.1
torchvision==0.4.2
traitlets==4.3.3
wcwidth==0.1.7
